requirements.txt
alabaster==0.7.10
APScheduler==3.4.0
attrs==17.3.0
Automat==0.6.0
Babel==2.5.1
backports.functools-lru-cache==1.4
backports.weakref==1.0.post1
bleach==1.5.0
boto3==1.4.7
botocore==1.7.43
certifi==2017.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.1
ConfigArgParse==0.12.0
configparser==3.5.0
constantly==15.1.0
cycler==0.10.0
cymem==1.31.2
cytoolz==0.8.2
decorator==4.1.2
dill==0.2.7.1
Django==1.11.7
django-jet==1.0.6
django-mptt==0.8.7
docopt==0.6.2
docutils==0.14
en-core-web-sm==2.0.0
entrypoints==0.2.3
enum34==1.1.6
fakeredis==0.9.0
Flask==0.12.2
ftfy==4.4.3
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
futures==3.1.1
gevent==1.2.2
graphviz==0.8.1
greenlet==0.4.12
h5py==2.7.1
html5lib==0.999999999
hyperlink==17.3.1
idna==2.6
imagesize==0.7.1
incremental==17.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
jsonpickle==0.9.5
jsonschema==2.6.0
jupyter-core==4.4.0
Keras==2.1.0
klein==17.10.0
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.1.0
mistune==0.8.1
mock==2.0.0
msgpack-numpy==0.4.1
msgpack-python==0.4.8
murmurhash==0.28.0
nbconvert==5.3.1
nbformat==4.4.0
nbsphinx==0.2.17
networkx==2.0
numpy==1.13.3
pandoc==1.0.2
pandocfilters==1.4.2
pathlib==1.0.1
pbr==3.1.1
plac==0.9.6
ply==3.10
preshed==1.0.0
protobuf==3.4.0
Pygments==2.2.0
pykwalify==1.6.0
pymessenger==0.0.7.0
pyparsing==2.2.0
python-crfsuite==0.9.5
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
rasa-core==0.7.4
rasa-nlu==0.10.4
redis==2.10.6
regex==2017.4.5
requests==2.18.4
requests-toolbelt==0.8.0
s3transfer==0.1.11
scikit-learn==0.19.1
scipy==1.0.0
simplejson==3.12.0
six==1.11.0
sklearn-crfsuite==0.3.6
snowballstemmer==1.2.1
spacy==2.0.2
Sphinx==1.6.5
sphinxcontrib-websupport==1.0.1
subprocess32==3.2.7
tabulate==0.8.1
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0rc2
termcolor==1.1.0
testpath==0.3.1
thinc==6.10.0
toolz==0.8.2
tqdm==4.19.4
traitlets==4.3.2
Twisted==17.9.0
typing==3.6.2
tzlocal==1.4
ujson==1.35
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
wheel==0.30.0
wrapt==1.10.11
zope.interface==4.4.3
